+++ plugins/math-renderer/shared/markdown-it.d.ts
@@ -1,4 +1,0 @@
-declare module "markdown-it/dist/markdown-it.js" {
-  import MarkdownIt from "markdown-it";
-  export default MarkdownIt;
-}